seguranca/requirements.txt
appdirs==1.4.3
Django==1.11
django-tastypie==0.13.3
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2017.2
six==1.10.0